moving files into correct folder
relates to #26
