+++ src/router.js
@@ -2,8 +2,6 @@ const http = require('http');
 const users = require('./static.js');
 const fs = require('fs');
 const pg = require('pg');
-const getData = require('./queries/getData.js');
-const postData = require('./queries/postData.js');
 
 const router = (request, response) => {
   const endpoint = request.url.split('/')[1];
@@ -19,30 +17,10 @@ const router = (request, response) => {
       }
     });
   } else if (endpoint === "users") {
-    // TASK 1: replace the 3 lines below below with your own function that gets data from your database
-    getData((err, res) => {
-        if (err) throw err;
-        let output = JSON.stringify(res);
-        response.writeHead(200,{
-          'content-type': 'application/json'
-        });
+    // Replace the 3 lines below below with your own function that gets data from your database
+    const output = JSON.stringify(users);
+    response.writeHead(200, {"Content-Type": "application/json"});
     response.end(output);
-  });
-} else if (endpoint === "create-user") {
-  let data = '';
-    request.on('data', function(chunk) {
-      data+=chunk;
-    });
-    request.on('end', () => {
-      const name = queryString.parse(data).name;
-      const location = queryString.parse(data).location;
-
-      postData(name, location, (err, res) => {
-        if (err) console.log(err);
-        console.log('RES is', res);
-      });
-
-    });
   } else {
     const fileName = request.url;
     const fileType = request.url.split(".")[1];
